refact: Modulariza a Impressão do Extrato em Txt e Aplica Tratamento de Exceções
